Add compile rule automation
Prior to this commit the Rust examples corresponding to each MISRA
rule were not incorporated into any "whole" which made it diffucult
to see the results of the compile behaviour of each. After this
commit the project is a proper crate and each MISRA rule Rust
example is a test implemented and run with the help of the
`compiletest_rs` crate. Use `cargo test` to view which tests
fail to compile as expected, "passing" tests are those that
fail to compile.